+++ src/components/custom/_signup-flat.tsx
@@ -6,8 +6,8 @@ import { Localize, localize } from 'components/localization'
 import { Container } from 'components/containers'
 import device from 'themes/device'
 // SVG
-import Facebook from 'images/svg/custom/facebook-blue.svg'
-import Google from 'images/svg/custom/google.svg'
+import Facebook from 'images/svg/components/component_facebook_blue.svg'
+import Google from 'images/svg/components/component_google.svg'
 
 type SignupFlatProps = {
     autofocus?: boolean
