+++ src/components/custom/other-platforms.tsx
@@ -15,8 +15,8 @@ import Cryptocurrencies from 'images/svg/components/component_cryptocurrencies_n
 import Commodities from 'images/svg/components/component_commodities_nav.svg'
 import Community from 'images/svg/menu/community.svg'
 import Contact from 'images/svg/menu/contact.svg'
-import DerivedFX from 'images/svg/custom/derived-fx.svg'
-import DerivX from 'images/svg/custom/deriv-x.svg'
+import DerivedFX from 'images/svg/components/component_derived_fx.svg'
+import DerivX from 'images/svg/components/component_deriv_x.svg'
 import { DerivEzLogoRed } from 'images/svg/deriv-ez'
 import { DbotIcon } from 'images/svg/dbot'
 import DMT5 from 'images/svg/dmt5/dmt5-icon.svg'
@@ -24,19 +24,19 @@ import BinaryBot from 'images/svg/binarybot-icon.svg'
 import DTrader from 'images/svg/dtrader/dtrader-icon.svg'
 import DerivGo from 'images/svg/deriv-go/deriv-go-icon.svg'
 import DerivLife from 'images/svg/menu/deriv-life.svg'
-import Forex from 'images/svg/custom/forex-nav.svg'
+import Forex from 'images/svg/components/component_forex_nav.svg'
 import Help from 'images/svg/menu/help-center.svg'
-import CFD from 'images/svg/custom/margin-trading-nav.svg'
-import Multipliers from 'images/svg/custom/multipliers-nav.svg'
-import Options from 'images/svg/custom/options-nav.svg'
+import CFD from 'images/svg/components/component_margin_trading_nav.svg'
+import Multipliers from 'images/svg/components/component_multipliers_nav.svg'
+import Options from 'images/svg/components/component_options_nav.svg'
 import Partner from 'images/svg/menu/partner.svg'
 import Payment from 'images/svg/menu/payment-methods.svg'
 import RegulatoryInfo from 'images/svg/menu/regulatory.svg'
 import Signals from 'images/svg/menu/signals.svg'
-import Smarttrader from 'images/svg/custom/smarttrader.svg'
-import StockIndices from 'images/svg/custom/stock-indices-nav.svg'
+import Smarttrader from 'images/svg/components/component_smart_trader.svg'
+import StockIndices from 'images/svg/components/component_stock_indices_nav.svg'
 import Story from 'images/svg/menu/story.svg'
-import TraderTool from 'images/svg/custom/trader-tool-nav.svg'
+import TraderTool from 'images/svg/components/component_trader_tool_nav.svg'
 
 type CardProps = {
     is_selected?: boolean
