+++ src/components/elements/off-canvas-menu.tsx
@@ -23,35 +23,35 @@ import Choose from 'images/svg/menu/choose.svg'
 import Commodities from 'images/svg/components/component_commodities_nav.svg'
 import Community from 'images/svg/menu/community.svg'
 import Contact from 'images/svg/menu/contact.svg'
-import DerivedFX from 'images/svg/custom/derived-fx.svg'
+import DerivedFX from 'images/svg/components/component_derived_fx.svg'
 import DerivLife from 'images/svg/menu/deriv-life.svg'
 import { DbotIcon } from 'images/svg/dbot'
 import BinaryBot from 'images/svg/binarybot-icon.svg'
 import API from 'images/svg/menu/developers.svg'
 import Diagonal from 'images/svg/elements/pink-right-diagonal.svg'
 import DMT5 from 'images/svg/dmt5/dmt5-icon.svg'
-import DerivX from 'images/svg/custom/deriv-x.svg'
+import DerivX from 'images/svg/components/component_deriv_x.svg'
 import { DerivEzLogoRed } from 'images/svg/deriv-ez'
 import DTrader from 'images/svg/dtrader/dtrader-icon.svg'
 import DerivGo from 'images/svg/deriv-go/deriv-go-icon.svg'
-import Forex from 'images/svg/custom/forex-nav.svg'
+import Forex from 'images/svg/components/component_forex_nav.svg'
 import Help from 'images/svg/menu/help-center.svg'
-import CFD from 'images/svg/custom/margin-trading-nav.svg'
-import Multipliers from 'images/svg/custom/multipliers-nav.svg'
-import Options from 'images/svg/custom/options-nav.svg'
+import CFD from 'images/svg/components/component_margin_trading_nav.svg'
+import Multipliers from 'images/svg/components/component_multipliers_nav.svg'
+import Options from 'images/svg/components/component_options_nav.svg'
 import Partner from 'images/svg/menu/partner.svg'
 import Payment from 'images/svg/menu/payment-methods.svg'
 import PaymentAgent from 'images/svg/menu/payment-agent.svg'
 import BugBounty from 'images/svg/menu/bug-bounty.svg'
 import Regulatory from 'images/svg/menu/regulatory.svg'
 import SecureTrading from 'images/svg/menu/secure-trading.svg'
-import Smarttrader from 'images/svg/custom/smarttrader.svg'
+import Smarttrader from 'images/svg/components/component_smart_trader.svg'
 import Status from 'images/svg/elements/status.svg'
-import StockIndices from 'images/svg/custom/stock-indices-nav.svg'
+import StockIndices from 'images/svg/components/component_stock_indices_nav.svg'
 import Cryptocurrencies from 'images/svg/components/component_cryptocurrencies_nav.svg'
 import Story from 'images/svg/menu/story.svg'
 import Terms from 'images/svg/menu/terms.svg'
-import Trade from 'images/svg/custom/trader-tool-nav.svg'
+import Trade from 'images/svg/components/component_trader_tool_nav.svg'
 import Signals from 'images/svg/menu/signals.svg'
 import useRegion from 'components/hooks/use-region'
