+++ src/features/pages/home/our-platforms/data.tsx
@@ -3,12 +3,12 @@ import { StaticImage } from 'gatsby-plugin-image'
 import { SmartPlatformItem } from './types'
 import DTraderIcon from 'images/svg/custom/rebranding/dtrader-icon.svg'
 import DMT5Icon from 'images/svg/custom/rebranding/dmt5-icon.svg'
-import DerivGOIcon from 'images/svg/custom/rebranding/deriv-go-icon.svg'
-import DerivXIcon from 'images/svg/custom/rebranding/deriv-x-icon.svg'
-import DBotIcon from 'images/svg/custom/rebranding/deriv-bot-icon.svg'
-import BinaryBotIcon from 'images/svg/custom/rebranding/binarybot-icon.svg'
+import DerivGOIcon from 'images/svg/deriv_go_icon.svg'
+import DerivXIcon from 'images/svg/deriv_x_icon.svg'
+import DBotIcon from 'images/svg/deriv_bot_icon.svg'
+import BinaryBotIcon from 'images/svg/binarybot-icon.svg'
 import SmartTraderIcon from 'images/svg/custom/rebranding/smarttrader-icon.svg'
-import DerivEzIcon from 'images/svg/custom/rebranding/deriv-ez-icon.svg'
+import DerivEzIcon from 'images/svg/deriv_ez_icon.svg'
 import Link from 'features/components/atoms/link'
 import {
     deriv_go_ios_url,
