+++ src/images/svg/deriv-ez/index.ts
@@ -4,8 +4,8 @@ export { default as DerivEzLogo } from './deriv-ez-logo.svg'
 export { default as DerivEzQR } from './derivez-qr.svg'
 export { default as Dmt5Icon } from './dmt5-icon.svg'
 export { default as DtIcon } from '../custom/rebranding/dtrader-icon.svg'
-export { default as DxIcon } from '../custom/rebranding/deriv-x-icon.svg'
+export { default as DxIcon } from '../deriv_x_icon.svg'
 export { default as EasyToStart } from './easy-to-start.svg'
 export { default as FavouriteAssets } from './favourite-assets.svg'
-export { default as DGoIcon } from '../custom/rebranding/deriv-go-icon.svg'
+export { default as DGoIcon } from '../deriv_go_icon.svg'
 export { default as TwentyFourSeven } from './twenty-four-seven.svg'
