+++ src/pages/landing/ebooks/components/_get-ebook.tsx
@@ -11,8 +11,8 @@ import { Flex } from 'components/containers'
 import AgreementLabel from 'components/custom/_agreement-label'
 import device from 'themes/device'
 import Apple from 'images/svg/components/component_apple.svg'
-import Facebook from 'images/svg/custom/facebook-blue.svg'
-import Google from 'images/svg/custom/google.svg'
+import Facebook from 'images/svg/components/component_facebook_blue.svg'
+import Google from 'images/svg/components/component_google.svg'
 import ViewEmailImage from 'images/common/sign-up/view-email.png'
 import apiManager from 'common/websocket'
 import Link from 'features/components/atoms/link'
