+++ src/pages/markets/components/helper/_available-platforms.tsx
@@ -7,11 +7,11 @@ import { Header, Text } from 'components/elements'
 import { Localize, LocalizedLink } from 'components/localization'
 import device from 'themes/device'
 import SmartTrader from 'images/svg/custom/rebranding/smarttrader-icon.svg'
-import DerivX from 'images/svg/custom/deriv-x.svg'
-import DBot from 'images/svg/custom/rebranding/deriv-bot-icon.svg'
+import DerivX from 'images/svg/components/component_deriv_x.svg'
+import DBot from 'images/svg/deriv_bot_icon.svg'
 import DMT5 from 'images/svg/custom/rebranding/dmt5-icon.svg'
 import DTrader from 'images/svg/custom/rebranding/dtrader-icon.svg'
-import DerivGo from 'images/svg/custom/rebranding/deriv-go-icon.svg'
+import DerivGo from 'images/svg/deriv_go_icon.svg'
 import DerivEZ from 'images/svg/dtrader/ez-icon.svg'
 import useRegion from 'components/hooks/use-region'
